Arreglo rutas, validacion de token admin y mejoras para update, register y login user

+++ src/handlers/usersHandler.js
@@ -1,3 +1,4 @@
+import { where } from "sequelize";
 import {
   getAllUsersController,
   getUserByIdController,
@@ -8,9 +9,19 @@ import {
   loginController,
   updateProfileController,
 } from "../controllers/usersController.js";
+import { Op } from "sequelize";
 import { User } from "../database/db.js";
 
 const getAllUsersHandler = async (req, res) => {
+  // Obtener el rol del usuario autenticado desde la solicitud
+  const userRole = req.user.role;
+  // // Verificar si el usuario autenticado tiene permiso para actualizar
+  if (userRole !== "admin") {
+    return res
+      .status(403)
+      .send({ message: "Unauthorized operation: User is not an admin" });
+  }
+
   try {
     const users = await getAllUsersController();
     res.status(200).send(users);
@@ -22,10 +33,6 @@ const getAllUsersHandler = async (req, res) => {
 const getUserByIdHandler = async (req, res) => {
   try {
     const { id } = req.params;
-    // Obtener el rol del usuario autenticado desde la solicitud
-   // const userRole = req.user.role;
-    //if (userRole !== "admin")
-      //return res.status(400).json({ error: "User Unauthorized" });
     const user = await getUserByIdController(id);
     if (!user) throw new Error("User not found");
     res.status(200).send(user);
@@ -44,9 +51,10 @@ const genereteAuth0User = async (req, res) => {
   }
 };
 
-
 const registerHandler = async (req, res) => {
   const { email, password, name } = req.body;
+const userExisting = await User.findOne({where: {email}});
+if (userExisting)  return res.status(400).json({ message: "User already exists" });
   try {
     const user = await registercontroller(email, password, name);
     res.status(200).json(user);
@@ -56,9 +64,19 @@ const registerHandler = async (req, res) => {
 };
 
 const loginHandler = async (req, res) => {
-  const { email, password } = req.body;
+  const { email, password, name } = req.body;
+  console.log(email,password,name);
+  let userExisting = null;
   try {
-    const userExisting = await User.findOne({ where: { email } });
+    if (email != undefined) {
+      userExisting = await User.findOne({
+        where: { email: email },
+      });
+    } else {
+      userExisting = await User.findOne({
+        where: { name: name },
+      });
+    }
     if (!userExisting) {
       return res.status(400).json({ error: "Email not found" });
     }
@@ -68,8 +86,6 @@ const loginHandler = async (req, res) => {
     const token = await loginController(userExisting, password);
     res.cookie("token", token, { httpOnly: true });
     res.json({ message: "Inicio de sesión exitoso", token });
-    // res.json(authenticatedUser);
-    // console.log("Ingreso exitoso");
   } catch (error) {
     res.status(400).json({ message: error.message });
   }
@@ -100,25 +116,31 @@ const updateUserHandler = async (req, res) => {
 
 const updateProfileHandler = async (req, res) => {
   try {
-    const { id } = req.params;
+    const { email, id } = req.params;
     const updatedFields = req.body;
-    const updatedUser = await updateProfileController(id, updatedFields);
+
+    const updatedUser = await updateProfileController(
+      id ? id : email,
+      updatedFields
+    );
     res.status(200).send(updatedUser);
   } catch (error) {
     res.status(500).send({ message: error.message });
   }
-}
+};
 
 const deleteUserHandler = async (req, res) => {
   try {
     const { id } = req.params;
     // Obtener el rol del usuario autenticado desde la solicitud
-    //  const userRole = req.user.role;
+    const userRole = req.user.role;
+    // // Verificar si el usuario autenticado tiene permiso para actualizar
+    if (userRole !== "admin") {
+      return res
+        .status(403)
+        .send({ message: "Unauthorized operation: User is not an admin" });
+    }
 
-    //  // Verificar si el usuario autenticado tiene permiso para actualizar
-    //  if (userRole !== "admin") {
-    //      return res.status(403).send({ message: "Unauthorized operation: User is not an admin" });
-    //  }
     const deletedUser = await deleteUserController(id);
     res.status(200).send(deletedUser);
   } catch (error) {
